split person tests up into multiple files
